+++ src/notification/menus/__init__.py
@@ -4,6 +4,8 @@ notification/menus — 按菜单条目拆分的 Mixin 模块
 每个文件对应一个 /command 菜单，通过 Mixin 继承到 NotificationService。
 """
 from ._base import ImportBaseMixin
+from .messages import MessagesMixin
+from .help import HelpMenuMixin
 from .search import SearchMenuMixin
 from .auto import AutoMenuMixin
 from .url import UrlMenuMixin
@@ -14,6 +16,8 @@ from .cache import CacheMenuMixin
 
 __all__ = [
     "ImportBaseMixin",
+    "MessagesMixin",
+    "HelpMenuMixin",
     "SearchMenuMixin",
     "AutoMenuMixin",
     "UrlMenuMixin",
