+++ src/services/notification_service.py
@@ -12,6 +12,8 @@ from typing import Any, Callable, Dict, List, Optional
 from src.notification.base import CommandResult, ConversationState, ChannelCapabilities
 from src.notification.menus import (
     ImportBaseMixin,
+    MessagesMixin,
+    HelpMenuMixin,
     SearchMenuMixin,
     AutoMenuMixin,
     UrlMenuMixin,
@@ -28,6 +30,8 @@ PAGE_SIZE = 5
 
 class NotificationService(
     ImportBaseMixin,
+    MessagesMixin,
+    HelpMenuMixin,
     SearchMenuMixin,
     AutoMenuMixin,
     UrlMenuMixin,
@@ -352,312 +356,3 @@ class NotificationService(
                     self._task_progress_tg_msg.setdefault(task_id, {})[ch_id] = msg_id_out[0]
             except Exception as e:
                 logger.debug(f"渠道 {ch_id} 发送任务进度通知失败: {e}")
-
-    # ═══════════════════════════════════════════
-    # 事件消息格式化
-    # ═══════════════════════════════════════════
-
-    EVENT_LABELS = {
-        "import_success": ("导入成功", True),
-        "import_failed": ("导入失败", False),
-        "refresh_success": ("刷新成功", True),
-        "refresh_failed": ("刷新失败", False),
-        "auto_import_success": ("自动导入成功", True),
-        "auto_import_failed": ("自动导入失败", False),
-        "webhook_triggered": ("Webhook 触发", True),
-        "webhook_import_success": ("Webhook 导入成功", True),
-        "webhook_import_failed": ("Webhook 导入失败", False),
-        "incremental_refresh_success": ("追更刷新成功", True),
-        "incremental_refresh_failed": ("追更刷新失败", False),
-        "media_scan_complete": ("媒体库扫描完成", True),
-        "scheduled_task_complete": ("定时任务完成", True),
-        "scheduled_task_failed": ("定时任务失败", False),
-        "system_start": ("系统启动", True),
-        # 后备下载任务
-        "download_fallback_success": ("后备弹幕下载完成", True),
-        "download_fallback_failed": ("后备弹幕下载失败", False),
-    }
-
-    def _format_event_message(self, event_type: str, data: Dict[str, Any]) -> tuple:
-        """根据事件类型格式化通知消息，返回 (title, text)"""
-        label_info = self.EVENT_LABELS.get(event_type)
-        if not label_info:
-            return (event_type, data.get("text", ""))
-
-        label, is_success = label_info
-        task_title  = data.get("task_title", "")
-        message     = data.get("message", "")
-        task_id     = data.get("task_id", "")
-        anime_title = data.get("anime_title", "")
-        season      = data.get("season")
-        episode     = data.get("episode")
-        ep_count    = data.get("episode_count")
-        search_term = data.get("search_term", "")
-        search_type = data.get("search_type", "")
-        webhook_src = data.get("webhook_source", "")
-        provider    = data.get("provider", "")
-        media_id    = data.get("media_id", "")
-        tmdb_id     = data.get("tmdb_id", "")
-        media_type  = data.get("media_type", "")
-        finished_at = data.get("finished_at", "")
-        icon        = "✅" if is_success else "❌"
-        status_str  = "处理完成" if is_success else "处理失败"
-        msg_short   = (message[:300] + "…") if len(message) > 300 else message
-
-        # ── 系统启动 ──────────────────────────────────────
-        if event_type == "system_start":
-            return (label, "弹幕服务器已启动完成 ✓")
-
-        # ── Webhook 触发 ───────────────────────────────────
-        if event_type == "webhook_triggered":
-            anime  = anime_title or "未知"
-            source = webhook_src
-            delayed    = data.get("delayed", False)
-            delay_hours = data.get("delay_hours", "")
-            lines = [
-                "📺 *媒体信息*",
-                f"• 名称: {anime}",
-                f"• 来源: {source}",
-                f"• 操作: {'⏳ 延迟入库 ' + str(delay_hours) + ' 小时后执行' if delayed else '⚡ 即时导入'}",
-            ]
-            return (label, "\n".join(lines))
-
-        # ── 后备弹幕下载 ──────────────────────────────────
-        if event_type in ("download_fallback_success", "download_fallback_failed"):
-            lines = [
-                "📺 *媒体信息*",
-                f"• 任务: {task_title}" if task_title else "",
-                "",
-                "⚙️ *执行结果*",
-                f"• 状态: {icon} {status_str}",
-                f"  └─ 📋 {msg_short}" if msg_short else "",
-                f"• 时间: {finished_at}" if finished_at else "",
-            ]
-            return (f"{icon} {label}", "\n".join(l for l in lines if l != "─skip─"))
-
-        # ── 媒体库扫描 ────────────────────────────────────
-        if event_type == "media_scan_complete":
-            lines = [
-                "⚙️ *执行结果*",
-                f"• 状态: {icon} 扫描完成",
-                f"  └─ 📋 {msg_short}" if msg_short else "",
-                f"• 时间: {finished_at}" if finished_at else "",
-                f"• TaskID: `{task_id[:8]}…`" if task_id else "",
-            ]
-            return (f"{icon} {label}", "\n".join(l for l in lines if l))
-
-        # ── 定时任务 ──────────────────────────────────────
-        if event_type in ("scheduled_task_complete", "scheduled_task_failed"):
-            # 支持多行消息：按换行拆分，使用树形符号
-            if msg_short:
-                msg_lines = [l for l in msg_short.splitlines() if l.strip()]
-                if len(msg_lines) <= 1:
-                    detail_lines = [f"  └─ 📋 {msg_short}"]
-                else:
-                    detail_lines = [f"  ├─ 📋 {l}" for l in msg_lines[:-1]]
-                    detail_lines.append(f"  └─ 📋 {msg_lines[-1]}")
-            else:
-                detail_lines = []
-            lines = [
-                "⚙️ *执行结果*",
-                f"• 任务: {task_title}" if task_title else "",
-                f"• 状态: {icon} {'已完成' if is_success else '执行失败'}",
-                *detail_lines,
-                f"• 时间: {finished_at}" if finished_at else "",
-                f"• TaskID: `{task_id[:8]}…`" if task_id else "",
-            ]
-            return (f"{icon} {label}", "\n".join(l for l in lines if l))
-
-        # ── 刷新类 ────────────────────────────────────────
-        if "refresh" in event_type:
-            s_str = f"S{int(season):02d}" if season is not None else ""
-            e_str = f"E{int(episode):02d}" if episode is not None else ""
-            lines = [
-                "📺 *媒体信息*",
-                f"• 名称: {anime_title}" if anime_title else "",
-                f"• 季集: {s_str}{e_str}" if s_str else "",
-                f"• 操作: 刷新弹幕",
-                f"• 状态: {icon} {status_str}",
-                f"• 更新集数: {ep_count} 集" if ep_count is not None else "",
-                "",
-                "⚙️ *执行结果*",
-                f"• TaskID: `{task_id[:8]}…`" if task_id else "",
-                f"  └─ 状态: {icon} 已完成 (100%)" if is_success else f"  └─ 状态: ❌ 失败",
-                f"  └─ 📋 {msg_short}" if msg_short else "",
-                f"• 时间: {finished_at}" if finished_at else "",
-            ]
-            return (f"{icon} {label}", "\n".join(l for l in lines if l))
-
-        # ── 导入类（import / auto_import / webhook_import）──
-        if "auto" in event_type:
-            header   = "🤖 自动导入通知"
-            op_label = "自动导入"
-        elif "webhook" in event_type:
-            header   = "🔗 Webhook 导入通知"
-            op_label = "Webhook 导入"
-        else:
-            header   = "🎬 任务导入通知"
-            op_label = "导入"
-
-        type_label_map = {"tv_series": "剧集", "movie": "电影", "tv": "剧集"}
-        media_type_zh  = type_label_map.get(media_type, media_type) if media_type else ""
-
-        provider_str = ""
-        if provider and (tmdb_id or media_id):
-            pid = tmdb_id or media_id
-            provider_str = f"{provider.upper()} {pid}"
-        elif provider:
-            provider_str = provider.upper()
-
-        s_str = f"S{int(season):02d}" if season is not None else ""
-        e_str = f"E{int(episode):02d}" if episode is not None else ""
-
-        # 操作描述
-        if s_str and e_str:
-            op_str = f"{op_label}{s_str}{e_str}"
-        elif s_str:
-            op_str = f"{op_label}{s_str}"
-        elif ep_count is not None:
-            op_str = f"{op_label}（共 {ep_count} 集）"
-        else:
-            op_str = op_label
-
-        type_map = {"keyword": "关键词", "tmdb": "TMDB", "tvdb": "TVDB",
-                    "douban": "豆瓣", "imdb": "IMDB", "bangumi": "Bangumi"}
-
-        lines = [
-            f"*{header}*",
-            "",
-            "📺 *媒体信息*",
-            f"• 名称: {anime_title}" if anime_title else (f"• 搜索词: {search_term}" if search_term else ""),
-            f"• 类型: {media_type_zh}" if media_type_zh else "",
-            f"• 操作: {op_str}",
-            f"• 状态: 🔄 入库中 → {icon} {status_str}",
-            f"• 季度: {s_str}" if s_str else "",
-            f"• 统计: {msg_short.split('，')[0]}" if msg_short and "新增" in msg_short else "",
-            f"• Provider: {provider_str}" if provider_str else "",
-            f"• 来源: {webhook_src}" if webhook_src else "",
-            f"• 时间: {finished_at}" if finished_at else "",
-            "",
-            "⚙️ *任务执行信息*",
-            f"• TaskID: `{task_id}`" if task_id else "",
-            f"  └─ 状态: {icon} 已完成 (100%)" if is_success else f"  └─ 状态: ❌ 失败",
-            f"  └─ 📋 {msg_short}" if msg_short else "",
-        ]
-        return (f"{icon} {label}", "\n".join(l for l in lines if l))
-
-    def _format_task_progress_message(self, task_title: str, progress: int, description: str) -> tuple:
-        """格式化任务进度消息，返回 (title, text)，适用于后备和普通下载任务"""
-        filled = int(progress / 10)
-        bar = "█" * filled + "░" * (10 - filled)
-        lines = ["⚙️ *执行进度*", ""]
-        if task_title:
-            lines.append(f"• 任务: {task_title}")
-        lines.append(f"• 进度: `[{bar}]` {progress}%")
-        if description:
-            lines.append(f"• 状态: {description}")
-        return ("⬇️ 任务进行中", "\n".join(lines))
-
-    # ═══════════════════════════════════════════
-    # 命令实现
-    # ═══════════════════════════════════════════
-
-    HELP_TEXT = (
-        "📖 *命令列表:*\n\n"
-        "🔍 /search <关键词> - 搜索弹幕源\n"
-        "  _支持指定季集，如: /search 刀剑神域 S01E10_\n"
-        "🔄 /auto - 自动导入（多平台）\n"
-        "🔗 /url - 从URL导入弹幕\n"
-        "♻️ /refresh - 弹幕库管理\n"
-        "🔑 /tokens - Token管理\n"
-        "📋 /tasks - 定时任务列表\n"
-        "🗑️ /cache - 清除缓存\n"
-        "❌ /cancel - 取消当前操作\n"
-        "📖 /help - 显示此帮助\n\n"
-        "💡 点击下方按钮可快速执行命令"
-    )
-
-    HELP_BUTTONS = [
-        [{"text": "🔍 搜索弹幕", "callback_data": "help_cmd:search"},
-         {"text": "🔄 自动导入", "callback_data": "help_cmd:auto"}],
-        [{"text": "🔗 URL导入", "callback_data": "help_cmd:url"},
-         {"text": "♻️ 弹幕库管理", "callback_data": "help_cmd:refresh"}],
-        [{"text": "🔑 Token管理", "callback_data": "help_cmd:tokens"},
-         {"text": "📋 任务列表", "callback_data": "help_cmd:tasks"}],
-        [{"text": "🗑️ 清除缓存", "callback_data": "help_cmd:cache"}],
-    ]
-
-    async def cmd_start(self, args: str, user_id: str, channel, **kw) -> CommandResult:
-        return CommandResult(
-            text=f"👋 欢迎使用弹幕服务器通知机器人！\n\n{self.HELP_TEXT}",
-            reply_markup=self.HELP_BUTTONS,
-            parse_mode="Markdown",
-        )
-
-    async def cmd_help(self, args: str, user_id: str, channel, **kw) -> CommandResult:
-        return CommandResult(
-            text=self.HELP_TEXT,
-            reply_markup=self.HELP_BUTTONS,
-            parse_mode="Markdown",
-        )
-
-    # ── /tasks ──
-
-    async def cmd_list_tasks(self, args: str, user_id: str, channel, **kw) -> CommandResult:
-        return await self._build_tasks_result()
-
-    async def _build_tasks_result(self, edit_message_id: int = None) -> CommandResult:
-        if not self.scheduler_manager:
-            return CommandResult(success=False, text="调度服务未就绪。")
-        try:
-            tasks = await self.scheduler_manager.get_all_tasks()
-            if not tasks:
-                return CommandResult(text="📋 当前没有定时任务。")
-            lines = ["📋 定时任务列表:\n"]
-            for t in tasks:
-                status = "✅" if t.get("isEnabled") else "⏸️"
-                name = t.get("name", "未知")
-                cron = t.get("cron", "")
-                lines.append(f"{status} {name} ({cron})")
-            return CommandResult(
-                text="\n".join(lines),
-                reply_markup=[[{"text": "🔄 刷新", "callback_data": "tasks_refresh"}]],
-                edit_message_id=edit_message_id,
-            )
-        except Exception as e:
-            logger.error(f"获取任务列表失败: {e}", exc_info=True)
-            return CommandResult(success=False, text=f"获取任务列表出错: {e}")
-
-    async def cb_tasks_refresh(self, params, user_id, channel, **kw):
-        msg_id = kw.get("message_id")
-        return await self._build_tasks_result(edit_message_id=msg_id)
-
-    async def cb_noop(self, params, user_id, channel, **kw):
-        return CommandResult(text="", answer_callback_text="")
-
-    async def cb_help_cmd(self, params, user_id, channel, **kw):
-        """帮助页内联按钮 — 点击后触发对应命令"""
-        cmd = params[0] if params else ""
-        handler_map = {
-            "search": self.cmd_search,
-            "auto": self.cmd_auto,
-            "url": self.cmd_url,
-            "refresh": self.cmd_refresh,
-            "tokens": self.cmd_list_tokens,
-            "tasks": self.cmd_list_tasks,
-            "cache": self.cmd_cache,
-        }
-        handler = handler_map.get(cmd)
-        if not handler:
-            return CommandResult(text="", answer_callback_text="未知命令")
-        self.clear_conversation(user_id)
-        result = await handler("", user_id, channel, **kw)
-        result.answer_callback_text = ""
-        result.edit_message_id = kw.get("message_id")
-        return result
-
-    # ═══════════════════════════════════════════
-    # 以下方法均已迁移到 src/notification/menus/ 下的 Mixin，
-    # 此处保留注释占位，便于 IDE 识别继承关系
-    # ═══════════════════════════════════════════
-    # （所有 cmd_*/cb_*/_{text/do/build/submit}* 方法已通过 Mixin 继承）
